refactor: move data protocols to it respective type folders
