--- src/data/protocols/db/delete-account-repository.ts
@@ -1,3 +1,0 @@
-export interface IDeleteAccountRepository {
-  delete(id: string): Promise<string>
-}
