--- src/data/protocols/db/get-all-tasks-repository.ts
@@ -1,5 +1,0 @@
-import { ITask } from "@domain/protocols/task";
-
-export interface IGetAllTasksRepository {
-  getAll(id: string): Promise<ITask[]>
-}
