+++ src/data/protocols/db/task/get-all-tasks-repository.ts
@@ -1,0 +1,5 @@
+import { ITask } from "@domain/protocols/task";
+
+export interface IGetAllTasksRepository {
+  getAll(id: string): Promise<ITask[]>
+}
