--- src/data/protocols/db/update-account-repository.ts
@@ -1,6 +1,0 @@
-import { IAccount } from "../../../domain/protocols/account";
-import { UpdateAccountModel } from "../../../domain/usecases/users/update-account";
-
-export interface IUpdateAccountRepository {
-  update(account: UpdateAccountModel): Promise<IAccount>
-}
