+++ src/data/protocols/db/users/add-account-repository.ts
@@ -1,0 +1,6 @@
+import { IAccount } from '../../../../domain/protocols/account'
+import { IAddAccountModel } from '../../../../domain/usecases/users/add-account'
+
+export interface IAddAccountRepository {
+  add(account: IAddAccountModel): Promise<IAccount>
+}
