+++ src/data/protocols/db/users/delete-account-repository.ts
@@ -1,0 +1,3 @@
+export interface IDeleteAccountRepository {
+  delete(id: string): Promise<string>
+}
