+++ src/data/usecases/db/tasks/get/db-get-all-tasks.ts
@@ -1,4 +1,4 @@
-import { IGetAllTasksRepository } from "@data/protocols/db/get-all-tasks-repository";
+import { IGetAllTasksRepository } from "@data/protocols/db/task/get-all-tasks-repository";
 import { ITask } from "@domain/protocols/task";
 import { IGetAllTasks } from "@domain/usecases/tasks/get-all-tasks";
