+++ src/data/usecases/db/users/add-account/db-add-account.spec.ts
@@ -1,5 +1,5 @@
 import { IHasher } from '../../../../protocols/criptography/hasher'
-import { IAddAccountRepository } from '../../../../protocols/db/add-account-repository'
+import { IAddAccountRepository } from '../../../../protocols/db/users/add-account-repository'
 import { IAddAccountModel } from '../../../../../domain/usecases/users/add-account'
 import { IAccount } from '../../../../../domain/protocols/account'
 import { DbAddAccount } from './db-add-account'
